Moved RemoteLogReadResult and RemoteStorageThreadPool to storage module
