+++ core/src/main/java/kafka/log/remote/RemoteLogReader.java
@@ -19,6 +19,7 @@ package kafka.log.remote;
 import org.apache.kafka.common.errors.OffsetOutOfRangeException;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.storage.internals.log.FetchDataInfo;
+import org.apache.kafka.storage.internals.log.RemoteLogReadResult;
 import org.apache.kafka.storage.internals.log.RemoteStorageFetchInfo;
 import org.slf4j.Logger;
