+++ core/src/test/java/kafka/log/remote/RemoteLogReaderTest.java
@@ -22,6 +22,7 @@ import org.apache.kafka.common.record.Records;
 import org.apache.kafka.server.log.remote.storage.RemoteStorageException;
 import org.apache.kafka.storage.internals.log.FetchDataInfo;
 import org.apache.kafka.storage.internals.log.LogOffsetMetadata;
+import org.apache.kafka.storage.internals.log.RemoteLogReadResult;
 import org.apache.kafka.storage.internals.log.RemoteStorageFetchInfo;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
